feat(spawner): use abstract unix sockets for stdio apps
This commit refactors the spawner to use abstract unix sockets for
stdio-based applications. This avoids creating socket files on the
filesystem for stdio mode.

The log messages have been improved to be more concise and accurate.

The tests have been updated to cover the new implementation and to be
more robust.