feat(app-env): add -listenAddr flag for standalone mode
This commit adds a -listenAddr flag to the app-env application.
If the -listenAddr flag is provided, the application will run as a
standalone server on the specified address. Otherwise, it will run as a
FastCGI application (either socket or stdin).

This provides a more robust way to select the application's mode,
improving upon the previous reliance on environment variables.